Return unsigned int from nplurals()
Fix type mismatch warnings with MSVC in some places after 87127dfd.

+++ src/catalog_po.cpp
@@ -1466,7 +1466,7 @@ bool POCatalog::DoSaveOnly(wxTextBuffer& f, wxTextFileType crlf)
     SaveMultiLines(f, pohdr);
     f.AddLine(wxEmptyString);
 
-    auto pluralsCount = std::max((int)GetPluralFormsCountPresentInItems(), GetPluralForms().nplurals());
+    auto pluralsCount = std::max(GetPluralFormsCountPresentInItems(), GetPluralForms().nplurals());
 
     for (auto& data_: m_items)
     {
